SSP-40
Mraihi Amine
3 hours ago
Edited

Now fast �� getting the status of the containers inside the python code periodically each minute.

And boom �����it works

There was an error due to the print function the solution was to add flush=True

By adding flush=True as an argument to the print function, the logs will be immediately flushed to the standard output, ensuring they are displayed without any delay.